% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{BuildCCInx}
\alias{BuildCCInx}
\title{Build cell-cell interaction edge list from scored cell-type statistics}
\usage{
BuildCCInx(GeneStatList, GeneMagnitude, GeneStatistic,
  Species = "hsapiens")
}
\arguments{
\item{GeneStatList}{A named list of dataframes. Each list element should
represent a cell type / cluster to be included in the interaction network,
and should be named accordingly. List elements should contain data frames
where each row is a gene with official gene symbols as row names. Variables
should be appropriately named statistics or annotations to be included in
the resulting node metadata. Variable names should be consistent between
list elements.}

\item{GeneMagnitude}{A character vector of length 1 representing the variable
name in the GeneStatList data frames carrying information on the magnitude
(and direction of the change) of expression for the node (gene) in each
cell type. This is either a measure of expression (generally mean
expression or detection rate) or a measure of change (signed log expression
ratio a.k.a. logFC).}

\item{GeneStatistic}{Optional. A character vector of length 1 representing
the variable name in the GeneStatList data frames carrying information on
the statistical significance of expression change. This is generally a
corrected p-value.}

\item{Species}{Default='hsapiens'. The species of the source data. One of
'hsapiens' or 'mmusculus'. Note that the ligand-receptor database was built
for human, and the mouse version is generated by homology mapping (only
using uniquely mapped homologues).}
}
\description{
Build cell-cell interaction edge list from scored cell-type statistics
}
