% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{BuildGeneStatList}
\alias{BuildGeneStatList}
\title{Build cluster-wise list of gene expression statistics from scRNAseq data}
\usage{
BuildGeneStatList(inD, cl, assayType = "", exponent = 2,
  pseudocount = 1)
}
\arguments{
\item{inD}{The input dataset. An object of class \code{\link[Seurat]{seurat}}
or \code{\link[SingleCellExperiment]{SingleCellExperiment}}. Other data
classes are not currently supported.
\href{https://github.com/BaderLab/scClustViz/issues}{Please submit requests
for other data objects here!}}

\item{cl}{a factor where each value is the cluster assignment for a cell
(column) in the input gene expression matrix.}

\item{assayType}{Default = "" (for Seurat v1/2). A length-one character
vector representing the assay slot in which the expression data is stored
in the input object. This is not required for Seurat v1 or v2 objects. See
\code{\link[scClustViz]{getExpr}} for details.}

\item{exponent}{Default = 2. A length-one numeric vector representing the
base of the log-normalized gene expression data to be processed. Generally
gene expression data is transformed into log2 space when normalizing (set
this to 2), though \code{Seurat} uses the natural log (set this to exp(1)).}

\item{pseudocount}{Default = 1. A length-one numeric vector representing the
pseudocount added to all log-normalized values in your input data. Most
methods use a pseudocount of 1 to eliminate log(0) errors.}
}
\description{
This function takes a \code{Seurat} or \code{SingleCellExperiment} object and
builds a list of dataframes containing gene expression statistics for all
genes of each cluster. This can be used as the input to
\code{\link{BuildCCInx}} for generating cell-cell interaction predictions
between cell-type clusters.
}
\references{
Mean gene expression calculations
  Innes BT and Bader GD. scClustViz – Single-cell RNAseq cluster
  assessment and visualization [version 2; peer review: 2 approved].
  F1000Research 2019, 7:1522 (\url{https://doi.org/10.12688/f1000research.16198.2})
}
\seealso{
\code{\link[scClustViz]{CalcCGS}}
}
