% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{BuildInx}
\alias{BuildInx}
\title{Build cell-cell interaction edge list from scored cell-type statistics}
\usage{
BuildInx(SCL, NodeScoreThreshold = 0, EdgeScoreThreshold = 0,
  species = "hsapiens")
}
\arguments{
\item{SCL}{The output of \code{\link{ScoreClustGenes}}.}

\item{NodeScoreThreshold}{Default=0. The minimum absolute score required to
include a node in the cell-cell interaction network. The default does not
filter any nodes.}

\item{EdgeScoreThreshold}{Default=0. The minimum absolute score required to
include an edge in the cell-cell interaction network. Absolute edge score
is the summed magnitude of node scores. The default does not filter any
edges.}

\item{species}{Default='hsapiens'. The species of the source data. One of
'hsapiens' or 'mmusculus'. Note that the ligand-receptor database was built
for human, and the mouse version is generated by homology mapping (only
using uniquely mapped homologues).}
}
\description{
This function uses the output of \code{\link{ScoreClustGenes}} to build a
predicted cell-cell interaction network. It returns a list of bipartite
graphs (as edge lists) for each combination of cell types.
}
