% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{BuildInxNode}
\alias{BuildInxNode}
\title{Write node metadata for predicted cell-cell interaction networks}
\usage{
BuildInxNode(INX, SCL)
}
\arguments{
\item{INX}{The output of \code{\link{BuildInx}}.}

\item{SCL}{The output of \code{\link{ScoreClustGenes}}.}
}
\description{
This function uses the output of \code{\link{BuildInx}} and
\code{\link{ScoreClustGenes}} to write node metadata tables for all predicted
cell-cell interaction bipartite graphs.
}
