% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{CalcExprScaled}
\alias{CalcExprScaled}
\title{Check input and score and scale gene expression.}
\usage{
CalcExprScaled(gdb, expr)
}
\arguments{
\item{gdb}{A data frame representing gene statistics from a cell type, where
each row is a gene with official gene symbols as row names. Variables
should be appropriately named statistics or annotations to be included in
the resulting node metadata.}

\item{expr}{A character vector of length 1 representing the variable name
in the GeneStatList data frames carrying information on the magnitude and
direction of the change of expression for the node (gene) in each cell
type. This is generally a signed logFC or gene expression ratio.}
}
\description{
This function takes differential expression gene statistics from scRNAseq
data representing a cell type as a dataframe, and assigns scaled scores to
the statistic of choice. This is used to rank nodes and edges by differential
expression when viewing the bipartite ligand-receptor plots.
}
