% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{PlotCCInx}
\alias{PlotCCInx}
\title{Plot bipartite graphs of predicted cell-cell interaction networks.}
\usage{
PlotCCInx(INX, INXNODE, comparisons, directions = c("LtoR", "RtoL",
  "RtoE", "EtoR", "LtoE", "EtoL"), topN, topPer, imageFileType = "pdf",
  imageFilePath = "./")
}
\arguments{
\item{INX}{The output of \code{\link{BuildInx}}.}

\item{INXNODE}{The output of \code{\link{BuildInxNode}}.}

\item{comparisons}{Optional, default is to plot all pairs of cell types. A
character vector representing the names of \code{inx} to plot.}

\item{directions}{Optional, default is to plot all combinations of ligands,
receptors, and ECM components. A character vector representing sets of
interactions to consider. Must be a subset of
\code{c('LtoR','RtoL','RtoE','EtoR','LtoE','EtoL')}. 'L' refers to ligands,
'R' to receptors, and 'E' to ECM components. The order is the same as that
in \code{comparisons} (the names of \code{inx}), such that 'RtoL' for
\code{inx[["A~B"]]} will plot the receptors from cell type A and their
ligands from cell type B.}

\item{topN}{Optional. Number of edges to include in the plot, ranked by
absolute edge score. Only one of \code{topN} and \code{topPer} can be
specified.}

\item{topPer}{Optional. Number of most differentially expressed edges to
include from each side of the comparison. Only one of \code{topN} and
\code{topPer} can be specified.}

\item{imageFileType}{Default="pdf". The file format for saved figures. One of
\code{"pdf"} (generated with \code{\link[grDevices]{cairo_pdf}}),
\code{"eps"} (generated with \code{\link[grDevices]{cairo_ps}}),
\code{"tiff"} (generated with \code{\link[grDevices]{tiff}}), or
\code{"png"} (generated with \code{\link[grDevices]{png}}).}

\item{imageFilePath}{Default is working directory. Character vector
indicating where to save image files.}
}
\description{
This function generates figures displaying bipartite graphs for predicted
cell-cell interaction networks between pairs of cell types.
}
