% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{prepBipartiteGraphSpread}
\alias{prepBipartiteGraphSpread}
\title{Internal Fx - Prep for plotting}
\usage{
prepBipartiteGraphSpread(compName, INX, INXNODE, direction, topN, topPer)
}
\description{
Internal Fx - Prep for plotting
}
