% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyInx.R
\name{ViewCCInx}
\alias{ViewCCInx}
\title{Run the Shiny app to interactively explore cell-cell interactions.}
\usage{
ViewCCInx(INX, imageFileType = "pdf", ...)
}
\arguments{
\item{INX}{The list containing edgelist and node metadata as generated by
\code{\link{BuildCCInx}}.}

\item{imageFileType}{Default="pdf". The file format for saved figures. One of
\code{"pdf"} (generated with \code{\link[grDevices]{cairo_pdf}}),
\code{"eps"} (generated with \code{\link[grDevices]{cairo_ps}}),
\code{"tiff"} (generated with \code{\link[grDevices]{tiff}}), or
\code{"png"} (generated with \code{\link[grDevices]{png}}).}

\item{...}{Named options that should be passed to the
\code{\link[shiny]{runApp}} call (these can be any of the following:
"port", "launch.browser", "host", "quiet", "display.mode" and "test.mode").}
}
\description{
Run the Shiny app to interactively explore cell-cell interactions.
}
