% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildInx.R
\name{BuildCCInx}
\alias{BuildCCInx}
\title{Build cell-cell interaction predictions between cell types}
\usage{
BuildCCInx(GeneStatList, GeneMagnitude = "MeanNormGeneExpr",
  GeneStatistic, Species = "hsapiens")
}
\arguments{
\item{GeneStatList}{A named list of dataframes. Each list element should
represent a cell type / cluster to be included in the interaction network,
and should be named accordingly. List elements should contain data frames
where each row is a gene with official gene symbols as row names. Variables
should be appropriately named statistics or annotations to be included in
the resulting node metadata. Variable names should be consistent between
list elements. The function \code{\link{BuildGeneStatList}} can be used to
generate this list from \code{Seurat} or \code{SingleCellExperiment}
objects when generating predictions not involving differential gene
expression.}

\item{GeneMagnitude}{Default = "MeanNormGeneExpr". A character vector of length 1
representing the variable name in the GeneStatList data frames carrying
information on the magnitude (and direction of the change) of expression
for the node (gene) in each cell type. This is either a measure of
expression (generally mean expression or detection rate) or a measure of
change (signed log expression ratio a.k.a. logFC). Default assumes
\code{GeneStatList} is output from \code{\link{BuildGeneStatList}}, and
uses mean normalized gene expression to weight nodes and edges.}

\item{GeneStatistic}{Optional. A character vector of length 1 representing
the variable name in the GeneStatList data frames carrying information on
the statistical significance of expression change. This is generally a
corrected p-value.}

\item{Species}{Default='hsapiens'. The species of the source data. One of
'hsapiens' or 'mmusculus'. Note that the ligand-receptor database was built
for human, and the mouse version is generated by homology mapping (only
using uniquely mapped homologues).}
}
\description{
This function takes a list of gene statistics per cluster to predict
cell-cell interactions between each cell-type (cluster). If the
\code{GeneStatistic} argument is provided, this function will assume the gene
statistics represent differential expression between experimental conditions,
and will weight the predicted interactions accordingly. Otherwise,
predictions will be weighted by expression magnitude per cell type. The
output of this function can be explored interactively with
\code{\link{ViewCCInx}}, or static figures can be generated with
\code{\link{PlotCCInx}}.
}
